fix: deduplicate insertions formatting code
