+++ packages/nextalign_cli/CMakeLists.txt
@@ -12,6 +12,7 @@ set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 set(CMAKE_INSTALL_MESSAGE LAZY)
 
 find_package(Boost 1.75.0 COMPONENTS headers REQUIRED)
+find_package(Microsoft.GSL 3.1.0 REQUIRED)
 find_package(TBB REQUIRED)
 find_package(cxxopts 2.2.1 REQUIRED)
 find_package(fmt 7.1.0 REQUIRED)
@@ -82,6 +83,7 @@ target_link_libraries(${PROJECT_NAME}
   PRIVATE
   ${STATIC_BUILD_FLAGS}
   Boost::headers
+  Microsoft.GSL::GSL
   TBB::tbb
   cxxopts::cxxopts
   fmt::fmt-header-only
